Example Done

+++ src/main/java/ca/jrvs/apps/twitter/example/dto/Financial.java
@@ -25,17 +25,17 @@ public class Financial {
     @JsonProperty("reportDate")
     private String reportDate;
     @JsonProperty("grossProfit")
-    private Integer grossProfit;
+    private long grossProfit;
     @JsonProperty("costOfRevenue")
-    private Integer costOfRevenue;
+    private long costOfRevenue;
     @JsonProperty("operatingRevenue")
-    private Integer operatingRevenue;
+    private long operatingRevenue;
     @JsonProperty("totalRevenue")
-    private Integer totalRevenue;
+    private long totalRevenue;
     @JsonProperty("operatingIncome")
-    private Integer operatingIncome;
+    private long operatingIncome;
     @JsonProperty("netIncome")
-    private Integer netIncome;
+    private long netIncome;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -50,62 +50,62 @@ public class Financial {
     }
 
     @JsonProperty("grossProfit")
-    public Integer getGrossProfit() {
+    public long getGrossProfit() {
         return grossProfit;
     }
 
     @JsonProperty("grossProfit")
-    public void setGrossProfit(Integer grossProfit) {
+    public void setGrossProfit(long grossProfit) {
         this.grossProfit = grossProfit;
     }
 
     @JsonProperty("costOfRevenue")
-    public Integer getCostOfRevenue() {
+    public long getCostOfRevenue() {
         return costOfRevenue;
     }
 
     @JsonProperty("costOfRevenue")
-    public void setCostOfRevenue(Integer costOfRevenue) {
+    public void setCostOfRevenue(long costOfRevenue) {
         this.costOfRevenue = costOfRevenue;
     }
 
     @JsonProperty("operatingRevenue")
-    public Integer getOperatingRevenue() {
+    public long getOperatingRevenue() {
         return operatingRevenue;
     }
 
     @JsonProperty("operatingRevenue")
-    public void setOperatingRevenue(Integer operatingRevenue) {
+    public void setOperatingRevenue(long operatingRevenue) {
         this.operatingRevenue = operatingRevenue;
     }
 
     @JsonProperty("totalRevenue")
-    public Integer getTotalRevenue() {
+    public long getTotalRevenue() {
         return totalRevenue;
     }
 
     @JsonProperty("totalRevenue")
-    public void setTotalRevenue(Integer totalRevenue) {
+    public void setTotalRevenue(long totalRevenue) {
         this.totalRevenue = totalRevenue;
     }
 
     @JsonProperty("operatingIncome")
-    public Integer getOperatingIncome() {
+    public long getOperatingIncome() {
         return operatingIncome;
     }
 
     @JsonProperty("operatingIncome")
-    public void setOperatingIncome(Integer operatingIncome) {
+    public void setOperatingIncome(long operatingIncome) {
         this.operatingIncome = operatingIncome;
     }
 
     @JsonProperty("netIncome")
-    public Integer getNetIncome() {
+    public long getNetIncome() {
         return netIncome;
     }
 
     @JsonProperty("netIncome")
-    public void setNetIncome(Integer netIncome) {
+    public void setNetIncome(long netIncome) {
         this.netIncome = netIncome;
     }
